Add generalized Annulus alongside CircularProgress
